test : files for test refactor
테스트 코드 리펙토링
